backend/btc_paper_trader/requirements-btc.txt
numpy>=1.24
pandas>=2.0
requests>=2.28
pyyaml>=6.0
scikit-learn>=1.3
